separating out shared functionality into a mixin for pileup and synteny to inherit separately with their own implementations
