+++ plugins/alignments/src/index.ts
@@ -43,6 +43,7 @@ export default class AlignmentsPlugin extends Plugin {
 export {
   linearPileupDisplayStateModelFactory,
   linearPileupDisplayConfigSchemaFactory,
+  SharedLinearPileupDisplayMixin,
 } from './LinearPileupDisplay'
 export { type LinearPileupDisplayModel } from './LinearPileupDisplay/model'
 export * as MismatchParser from './MismatchParser'
